pyproject.toml
[project]
name = "Feed-Fusion"
version = "0.1.0"
description = ""
authors = [
    {name = "vavlxxx",email = "vladimir.nefedov357@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi[standart] (>=0.116.2,<0.117.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "alembic-postgresql-enum (>=1.8.0,<2.0.0)",
    "orjson (>=3.11.3,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "telethon (>=1.41.2,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=7.0.1,<8.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "pydantic[email] (>=2.4.1,<2.12)",
    "aiogram (>=3.22.0,<4.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "elasticsearch (==8.11.0)",
    "celery-types (>=0.24.0,<0.25.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "torch (>=2.10.0,<3.0.0)",
]

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-cpu"}

[tool.poetry]
package-mode=false

[tool.ruff]
line-length = 68

[tool.pyright]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.13.0"
pyright = "^1.1.405"
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
